+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Prop to set whether or not to show the collection badges.
+
+### Fixed
+- Error when `productClusters` were `undefined`.
 
 ## [0.8.0] - 2018-6-20
 ### Added